% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/the pack v1.r
\name{datex.create_data}
\alias{datex.create_data}
\title{A Cat Function}
\usage{
datex.create_data()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}
